hot fix

+++ backend/api/views.py
@@ -57,7 +57,7 @@ def put_data(request):
         try:
             leaderboard = api.models.LeaderboardModel.objects.filter(level=serializer.data["level"]).get(gamer=gamer)
             score_time = serializer.data["score"]
-            if int(leaderboard.score) > int(score_time):
+            if float(leaderboard.score) > float(score_time):
                 leaderboard.score = serializer.data["score"]
                 leaderboard.save()
         except api.models.LeaderboardModel.DoesNotExist:
